refactor: move assert_timestamps_in_range function to utils for better organization
