+++ tests/test_data_feed.py
@@ -1,5 +1,7 @@
 import pytest
 import numpy as np
+from .utils import assert_timestamps_in_range
+
 from ccxt_backtesting_exchange.data_feed import DataFeed
 
 
@@ -11,31 +13,6 @@ def data_feed():
 @pytest.fixture()
 def empty_data_feed():
     return DataFeed("./data/empty.json")
-
-
-def assert_timestamps_in_range(
-    data: np.ndarray,
-    first_expected_timestamp: np.float32,
-    last_expected_timestamp: np.float32,
-):
-    """
-    Assert that all timestamps in the data are within the specified range.
-
-    :param data: A NumPy array or list of timestamps.
-    :param first_expected_timestamp: The first timestamp returned.
-    :param last_expected_timestamp: The last allowable timestamp.
-    """
-    assert isinstance(data, np.ndarray), "Data must be a NumPy array"
-    assert len(data) > 0, "Data should not be empty"
-
-    returned_timestamps = data[:, 0]
-    first_returned_timestamp = returned_timestamps[0]
-    last_returned_timestamp = returned_timestamps[-1]
-
-    assert first_returned_timestamp == first_expected_timestamp
-    assert last_returned_timestamp == last_expected_timestamp
-    assert np.all(returned_timestamps >= first_expected_timestamp)
-    assert np.all(returned_timestamps <= last_expected_timestamp)
 
 
 def test_get_data_without_limit_returns_all(data_feed):
